Continued switch over to inttypes.h types.

+++ src/interpretation.h
@@ -117,14 +117,14 @@ class Flag
 {
 public:
     // TODO: Use different term than "interpretation", which is already used to mean something else
-    Flag(unsigned int numBits, initializer_list<string> flagValues);
-    Flag(unsigned int numBits, string flagValue); // Sets all bit sequences to the same meaning -- commonly used for "unused" or "reserved" bits
+    Flag(uint8_t numBits, initializer_list<string> flagValues);
+    Flag(uint8_t numBits, string flagValue); // Sets all bit sequences to the same meaning -- commonly used for "unused" or "reserved" bits
 
-    unsigned int getNumBits();
+    uint8_t getNumBits();
     string getInterpretation(unsigned int value);
 
 private:
-    unsigned int numBits;
+    uint8_t numBits;
     vector<string> flagValues;
 };
